requirements.txt
flask>=3.1.0
rich>=13.7.0
Pillow>=10.0.0